agv_backend/requirements.txt
asgiref==3.8.1
Django==4.2.23
django-cors-headers==4.7.0
django-restframework==0.0.1
djangorestframework==3.14.0
pytz==2025.2
sqlparse==0.5.3
typing_extensions==4.14.0
